pyproject.toml
[project]
name = "btwin-core"
version = "0.1.0"
description = "Core domain primitives for the B-TWIN collaboration control plane"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.6",
    "litellm>=1.55",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/btwin_core"]

[tool.hatch.build.targets.sdist]
include = [
    "src/btwin_core/global/providers.json",
    "src/btwin_core/global/protocols/**",
]
